CS: fix formatting of embedded PHP
Now, there are a couple of things fixes in one go here.

To start, the phrase contains placeholder and therefore needs a translators comment.
To add the translators comment, the embedded PHP would become multi-line (well, it already was, but shouldn't have been).
When embedded PHP is multi-line, the PHP open and close tags should each be on their own line without any other content.
And as the snippet now has become a multi-line snippet anyway, we may as well make the function call more readable by making it a proper multi-line function call.

Lastly, the phrase contained a new line character. These are meaningless in .po files and confusing for translators - should this be a `<br>` ?, does the format matter ? -, so this new line character has been removed.