CS/XSS: always escape output [3]
When a variable is used as part of an attribute or url, it is always better to escape the whole string as that way a potential escape character just before the variable will be correctly escaped.
Without the context of the complete string, - even when the variable is escaped - it may still leave you open to security issues.