Travis: actually run the unit tests
Why have the builds if we're not running anything on it ?
So, don't just run the unit tests for the code climate build, but on all the builds.

To allow the unit tests to run, a `composer install` needs to be done as we need the composer autoload file so the plugin class files can be found.

Also account for the fact that the PHPUnit version on the Travis images for PHP 7.2 and higher is incompatible with the WP unit test suite.
In those cases, install PHPUnit 5.7 via Composer and run PHPUnit from the `vendor` directory (including for the code climate build).