Composer/PHPUnit: be explicit about supported PHPUnit versions
The unit tests for the Clicky plugin, in part, depend on the WordPress native unit test setup and loads the WP native `bootstrap.php` file to that end.

This means that the unit tests for the plugin are compatible with the same PHPUnit versions as the WP native unit tests, which at this time is `^3.6.12 || ^4.5 || ^5.7 || ^6.0 || ^7.0`.
While PHPUnit 3.x is supported for testing on PHP 5.2, it is not listed in the `composer.json` file as Composer itself has a minimum requirement of PHP 5.3, so listing it is a moot point.

And as PHPUnit is now included in the `composer.json` and the `composer.lock` file, there is no need anymore for the explicit `require` when using a PHP > 7.1 image in the Travis script.