QA: Unit tests: use a more appropriate assertion
`assertContains()` can be used to examine both arrays as well as strings.
When passed a string as the `$haystack`, it will examine it to see if the `$needle` exists in the string.

So, basically that function does the same as the current test was doing, but then using PHPUnit native assertions,

Ref:
* https://phpunit.de/manual/6.5/en/appendixes.assertions.html#appendixes.assertions.assertContains#appendixes.assertions.assertContains.example2