Travis: fix test failures not being reported
Travis does not fail the build when there are multiple commands within a condition.

This means that, in this case, if the unit tests on the PHP 7.3 build would fail, Travis would still show a passing build.

Removing the `code_climate.sh` script and moving the Codeclimate reporting to the `after_script` section of the Travis script fixes this.