composer.lock: REVERT updates since addition of PHPUnit dependency
The update of the `composer.lock` file done in 42d1179b17dcc0c6fdacb624fb2ec376cbc07866 was run on PHP 7.2 or higher. This is causing issues as PHP 5.6 still needs to be supported by the Clicky plugin.

This commit reverts the `composer.lock` file to the state _before_ the faulty update.