Add .gitattributes file
This adds a `.gitattributes` file to keep the archives GH creates of the repo clean of development related files.
In effect, this means that the stable releases as available from GH as well as through Packagist will no longer contain the unit tests or other dev-only files.

People using Composer can still get these files in their setup if they really want to, by using `--prefer-source`.

The file list is based on the files in this repo, however a critical look at this list is warranted.

Refs:
* [Reddit: I don't need your tests in my production](https://www.reddit.com/r/PHP/comments/2jzp6k/i_dont_need_your_tests_in_my_production)
* [Blog: I don't need your tests in my production](https://blog.madewithlove.be/post/gitattributes/)