Travis: run the tests against PHP 8/nightly
As the tests for the Clicky plugin don't use mocking, the tests will run fine with PHPUnit 7.x on PHP 8.

All that is needed is to `ignore-platform-reqs` and as this repo has a committed `composer.lock` file, we also force installation of PHPUnit 7.x.

Includes adding `travis_retry` to `composer install`-like commands to get round builds stalling on the connection with Packagist.

Includes fixing the `composer test` command for cross-platform (i.e. Windows) compatibility.