Tests: modernize used assertions
The PHPUnit Polyfills allows for using PHPUnit 9.x assertions, even when the tests are being run on older PHPUnit versions.

This updates select assertions to the modern PHPUnit syntax, most notably, it switches out the following assertions:
* `assertContains()` with string haystacks for  `assertStringContainsString()`.

Note: the method calls switched over are based on tests run on PHPUnit 8/9 (don't ask how I did that) and only those assertions which _need_ to be switched over to prevent issues in the future, have been.