Travis: update for using correct PHPUnit version with WP 5.9+
As of WP 5.9, the integration tests can run on PHPUnit 5.7 - 9.x and should use the most appropriate PHPUnit version for the PHP version on which the tests are being run.

For WP < 5.9, the range of supported PHPUnit versions is still PHPUnit 5.7 - 7.x.

With that in mind and with there being a) a `config - platform - php` setting and b) a committed `composer.lock` file, some version juggling needs to be done.

For WP < 5.9, the logic previously used remains in place.
For WP 5.9, new logic is introduced.

Includes:
* Adding the PHPUnit cache file which is generated when using PHPUnit 8/9 to `.gitignore`.